Reapproach i18njs due to XSS concerns:  - A vendor script SHOULD NOT be responsible for parsing messages into the DOM  - mw.hook() should not be used to signal readiness for i18nJs, since anybody could use mw.hook() to fire any object
