+++ gadgets/AjaxBatchRedirect/code.js
@@ -384,8 +384,8 @@ maxerr: 999999, forin: false, -W082, -W084
   }
   mw.hook("dev.modal").add(preload);
   mw.hook('dev.powertools.placement').add(preload);
-  loadMessages().then(function (messages) {
-    i18n = messages;
+  getI18nLoader().then(function (loader) {
+    i18n = prepareI18n(loader);
     preload();
   });
