+++ gadgets/HTMLAudioPlayer5/code.js
@@ -195,8 +195,9 @@
   /* AUTO-GENERATE BOILERPLATE LOGIC ON COMPILATION */
   INJECT_FANDOM_UTILS_I18N();
 
-  loadMessages().then(function (i18no) {
-    msg = i18no.msg;
+  getI18nLoader().then(function (i18nLoader) {
+    var i18n = prepareI18n(i18nLoader);
+    msg = i18n.msg.bind(i18n);
     mw.hook('wikipage.content').add(init);
   });
