+++ gadgets/gadgets-definition.yaml
@@ -30,7 +30,7 @@ gadgets:
     requires: 
       - FandoomUiUtilsModal
       - PowertoolsPlacement
-      - FandoomUtilsI18njs
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     styles:
@@ -50,7 +50,7 @@ gadgets:
     requires: 
       - FandoomUiUtilsModal
       - PowertoolsPlacement
-      - FandoomUtilsI18njs
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     styles:
@@ -68,7 +68,7 @@ gadgets:
     requires: 
       - FandoomUiUtilsModal
       - PowertoolsPlacement
-      - FandoomUtilsI18njs
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     styles:
@@ -85,6 +85,8 @@ gadgets:
       - CoolMikeHatsune22 (port)
     links:
       - https://dev.fandom.com/wiki/AjaxUndo
+    requires:
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     i18n:
@@ -101,7 +103,7 @@ gadgets:
     links:
       - https://dev.fandom.com/wiki/Countdown
     requires:
-      - FandoomUtilsI18njs
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     i18n:
@@ -148,7 +150,7 @@ gadgets:
     scripts:
       - code.js
 
-  FandoomUtilsI18njs:
+  FandoomUtilsI18nLoader:
     description: "Library for loading i18n messages."
     authors:
       - Cqm
@@ -157,7 +159,11 @@ gadgets:
     links:
       - https://dev.fandom.com/wiki/I18n-js
     scripts:
-      - code.js
+      - gadget-impl.js
+    resourceLoader:
+      dependencies:
+        - mediawiki.jqueryMsg
+    disableBuildingOfGadgetImplJs: true
   
   HTMLAudioPlayer5:
     description: "Play audio files with a native HTML5 player"
@@ -168,6 +174,8 @@ gadgets:
       - CoolMikeHatsune22 (port)
     links:
       - https://dev.fandom.com/wiki/HTML5AudioPlayer
+    requires:
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     i18n:
@@ -185,7 +193,7 @@ gadgets:
       - FandoomUiUtilsModal
       - FandoomUiUtilsDorui
       - PowertoolsPlacement
-      - FandoomUtilsI18njs
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     styles:
@@ -214,7 +222,7 @@ gadgets:
     requires:
       - FandoomUiUtilsQdmodal
       - PowertoolsPlacement
-      - FandoomUtilsI18njs
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     styles:
@@ -233,7 +241,7 @@ gadgets:
     requires:
       - FandoomUiUtilsModal
       - PowertoolsPlacement
-      - FandoomUtilsI18njs
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     styles:
@@ -252,7 +260,7 @@ gadgets:
     requires:
       - FandoomUiUtilsModal
       - PowertoolsPlacement
-      - FandoomUtilsI18njs
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     styles:
@@ -269,6 +277,8 @@ gadgets:
       - CoolMikeHatsune22 (port)
     links:
       - https://dev.fandom.com/wiki/PageRenameAuto-update
+    requires:
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     styles:
@@ -295,7 +305,7 @@ gadgets:
     links:
       - https://dev.fandom.com/wiki/PreloadTemplates
     requires:
-      - FandoomUtilsI18njs
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
     styles:
@@ -313,7 +323,7 @@ gadgets:
     links:
       - https://dev.fandom.com/wiki/View_Source
     requires:
-      - FandoomUtilsI18njs
+      - FandoomUtilsI18nLoader
     scripts:
       - code.js
       - definitions.js
